requirements.txt
line-bot-sdk==1.15.0
Flask==2.1.0
requests==2.12.4